backend/requirements.txt
# Core
Flask>=3.0.0,<4
python-dotenv>=1.0.0

# Persistence
Flask-SQLAlchemy>=3.1.0,<4
SQLAlchemy>=2.0.0,<3
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Auth
Flask-JWT-Extended>=4.6.0,<5

# Validation
marshmallow>=3.21.0,<4
flask-marshmallow>=1.2.0,<2

# Cache & async jobs
redis>=5.0.0,<6
Flask-Caching>=2.3.0,<3
celery[redis]>=5.4.0,<6

# Migrations
Flask-Migrate>=4.0.0,<5

# CORS (frontend integration)
flask-cors>=4.0.0,<6

# Abuse protection
Flask-Limiter[redis]>=3.8.0,<4

# Production server
gunicorn>=22.0.0,<24
